refactor(d3): extract shared render helpers to utils/d3-helpers (109.2 step 2)
Move renderChartTitle, initD3Chart, createTooltip and the export-container
lifecycle helpers (EXPORT_WIDTH/HEIGHT, resolveExportPalette,
createExportContainer, finalizeSvgExport) out of d3.ts into utils/d3-helpers
so the per-viz renderer modules can share them. Pure move; zero drift.
